(feat) add Firestore SSR features to nightly tests (#24)
**General Updates:**
Updates the CI to test against Firebase Next and not Firebase Canary. The Canary target installs the Canary of the previously released version, where as Next installs the upcoming version (which we should test against).

Unfreezes the yarn lock file so that CI installs the latest Next version as it becomes available.

**Additions to the test framework:**
Expands upon Firestore tests to ensure that.

`toJson` and `fromJson` works for:
 * `Bytes`
 * `DocumentSnapshot`
 * `GeoPoint`
 * `QuerySnapshot`
 * `Timestamp`
 * `VectorValue`

Additionally tests that these object type can be serialized server-side then deserialized client-side and their values matched the originals.

Finally, also tests that QuerySnapshot and DocumentSnapshot can be stood up as listeners form their serialized forms.